requirements.txt
flask==2.3.3
flask-socketio==5.3.6
websockets==11.0.3
openai-whisper==20231117
faster-whisper==0.10.0
argostranslate==1.9.0
pydub==0.25.1
transformers==4.35.2
torch==2.1.0
torchaudio==2.1.0
numpy==1.24.3
requests==2.31.0
python-socketio==5.9.0
eventlet==0.33.3